+++ lca-core/src/shaders/elementwise_mul.wgsl
@@ -1,8 +1,8 @@
 // WGSL Shader to compute element-wise multiplication z = x * y
 
-@group(0) @binding(0) var<storage, read> x: array<f32>;
-@group(0) @binding(1) var<storage, read> y: array<f32>;
-@group(0) @binding(2) var<storage, read_write> z: array<f32>;
+@group(0) @binding(0) var<storage, read> x: array<f64>;
+@group(0) @binding(1) var<storage, read> y: array<f64>;
+@group(0) @binding(2) var<storage, read_write> z: array<f64>;
 
 @compute @workgroup_size(64)
 fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
