+++ lca-core/src/shaders/extract_diagonal.wgsl
@@ -3,10 +3,10 @@
 // Bindings for CSR matrix data
 @group(0) @binding(0) var<storage, read> row_pointers: array<u32>; // CSR row pointers (size N+1)
 @group(0) @binding(1) var<storage, read> col_indices: array<u32>; // CSR column indices (size NNZ)
-@group(0) @binding(2) var<storage, read> values: array<f32>;      // CSR values (size NNZ)
+@group(0) @binding(2) var<storage, read> values: array<f64>;      // CSR values (size NNZ)
 
 // Binding for the output diagonal vector
-@group(0) @binding(3) var<storage, read_write> diagonal_output: array<f32>; // Output vector (size N)
+@group(0) @binding(3) var<storage, read_write> diagonal_output: array<f64>; // Output vector (size N)
 
 // We could pass N via uniforms, but reading array_length is often simpler if possible,
 // or assume the dispatch size matches N. Let's assume dispatch size matches N.
@@ -26,8 +26,8 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let row_end = row_pointers[row_index + 1u];
 
     var found_diagonal = false;
-    // Explicitly declare diagonal_value as f32
-    var diagonal_value: f32 = 0.0;
+    // Explicitly declare diagonal_value as f64
+    var diagonal_value: f64 = 0.0;
 
     // Iterate through the non-zero elements of the current row
     for (var j = row_start; j < row_end; j = j + 1u) {
